Airtableコンテンツ用のコンポーネントをリファクタリング
