+++ src/components/contents/AppWriting/AppWriting.tsx
@@ -1,9 +1,8 @@
 import React, { VFC } from 'react'
 import styled from 'styled-components'
 import { graphql, useStaticQuery } from 'gatsby'
-import { Text } from 'smarthr-ui'
+import { TextAnchor } from '../shared/TextAnchor'
 import { FragmentTitle } from '../../article/FragmentTitle/FragmentTitle'
-import reactStringReplace from 'react-string-replace'
 import { marked } from 'marked'
 
 const query = graphql`
@@ -36,57 +35,43 @@ export const AppWriting: VFC = () => {
       recordId: node.data?.record_id,
       order: node.data?.order || Number.MAX_SAFE_INTEGER,
     }))
+    .filter((item) => {
+      return item.name
+    })
     .sort((x, y) => (x.order && y.order ? x.order - y.order : -1))
 
-  const getReplaceLinkText = (text: string) => {
-    return (
-      <Text as="p" whiteSpace="pre-wrap">
-        {reactStringReplace(text, /(https?:\/\/\S+)/g, (match: string, strIndex: number) => {
-          return (
-            <a key={strIndex} href={match}>
-              {match}
-            </a>
-          )
-        })}
-      </Text>
-    )
-  }
   return (
     <>
-      {appWritingData
-        .filter((item) => {
-          return item.name
-        })
-        .map(({ name, description, discussion, source, recordId }, index) => {
-          const fragmentId = (suffixId: string) => {
-            return recordId ? `${recordId}-${suffixId}` : ''
-          }
-          return (
-            <Wrapper key={index}>
-              <FragmentTitle tag="h2" id={fragmentId('0')}>
-                {name}
+      {appWritingData.map(({ name, description, discussion, source, recordId }, index) => {
+        const generateFragmentId = (suffixId: string) => {
+          return recordId ? `${recordId}-${suffixId}` : `${index}-${suffixId}`
+        }
+        return (
+          <Wrapper key={index}>
+            <FragmentTitle tag="h2" id={generateFragmentId('0')}>
+              {name}
+            </FragmentTitle>
+            {description && (
+              <FragmentTitle tag="h3" id={generateFragmentId('1')}>
+                説明
               </FragmentTitle>
-              {description && (
-                <FragmentTitle tag="h3" id={fragmentId('1')}>
-                  説明
-                </FragmentTitle>
-              )}
-              {description && <div dangerouslySetInnerHTML={{ __html: description }} />}
-              {discussion && (
-                <FragmentTitle tag="h3" id={fragmentId('2')}>
-                  議事録
-                </FragmentTitle>
-              )}
-              {discussion && getReplaceLinkText(discussion)}
-              {source && (
-                <FragmentTitle tag="h3" id={fragmentId('3')}>
-                  出典
-                </FragmentTitle>
-              )}
-              {source && getReplaceLinkText(source)}
-            </Wrapper>
-          )
-        })}
+            )}
+            {description && <div dangerouslySetInnerHTML={{ __html: description }} />}
+            {discussion && (
+              <FragmentTitle tag="h3" id={generateFragmentId('2')}>
+                議事録
+              </FragmentTitle>
+            )}
+            {discussion && <TextAnchor text={discussion} />}
+            {source && (
+              <FragmentTitle tag="h3" id={generateFragmentId('3')}>
+                出典
+              </FragmentTitle>
+            )}
+            {source && <TextAnchor text={source} />}
+          </Wrapper>
+        )
+      })}
     </>
   )
 }
